Code factorized.
