+++ src/main/java/com/project/airbnb_app/service/AppUserService.java
@@ -4,5 +4,10 @@ import com.project.airbnb_app.entity.User;
 
 public interface AppUserService {
 
+    /**
+     * Save user.
+     *
+     * @param user object.
+     */
     void save(User user);
 }
