+++ src/main/java/com/project/airbnb_app/service/HotelBookingService.java
@@ -4,8 +4,6 @@ import com.project.airbnb_app.entity.HotelBooking;
 
 public interface HotelBookingService {
 
-    HotelBooking findByPaymentSessionId(String paymentSessionId);
-
     String getBookingStatus(Long bookingId);
 
     void saveBooking(HotelBooking hotelBooking);
